ad74416h: various example fixes for STM32 platform
1. Move UART initialization from basic example to main.c to support
other examples.
2. Move set channel function to after DAC code setting in
current_input_loop example to avoid irrelevant short-circuit
detection.
3. Fix typo in temperature_2wire_rtd example.
4. Fix various variable types that break build due to
[-Wincompatible-pointer-types] flag.

+++ projects/ad74416h/src/examples/voltage_output/voltage_output.c
@@ -49,7 +49,7 @@ int example_main()
 	int ret;
 
 	uint16_t reg_value = 0xAAAA;
-	uint32_t dac_code = 0;
+	uint16_t dac_code = 0;
 
 	ret = ad74416h_init(&ad74416h_desc, &ad74416h_ip);
 	if (ret)
